refactor: 식당 상세 정보 조회 기능을 RestaurantReviewStatService로 이동
식당 조회에는 평균 별점과 리뷰 개수라는 집계 데이터가 포함되므로, 해당 데이터를 보관하고 있는 READ-DB의 RestaurantReviewStat 엔티티를 조회하도록 함.
